fix: move request structs to api folder
